requirements.txt
alembic==1.13.1
asyncpg==0.29.0
fastapi==0.109.0
pydantic==2.5.3
python-dotenv==1.0.0
SQLAlchemy==2.0.25
uvicorn==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
aioredis==2.0.1
celery==5.3.6
google-api-python-client==2.117.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread-asyncio==2.0.0
